shahul Update Selenium.md
The purpose of this automation test is to ensure that the search functionality on the Selenium Playground website works as intended. Automated testing is crucial in software development as it helps identify bugs and issues early in the development cycle, ensuring a smoother user experience.

## Requirements
To run this script, you will need the following:
Python 3.x: The programming language used for writing the automation script.
-Selenium A powerful tool for controlling web browsers through programs and performing browser automation.
Chrome WebDriver: A driver that allows Selenium to interact with the Chrome browser. You can also use other browser drivers  if preferred.